requirements.txt
streamlit==1.19.0
numpy==1.22.4
cv2==4.7.0
tensorflow==2.6.0
Pillow==9.4.0
joblib==1.2.0
scikit-learn==1.2.2
